pyproject.toml
[project]
name = "school-pro-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.24",
]
